ref: extracting agentic stuff into a dedicated file.
